feat: Detail Screen 덕퀴즈인지, 일반 시험인지에 따라 분기할 수 있도록 재정비
