+++ feature-ui-detail/src/main/kotlin/team/duckie/app/android/feature/ui/detail/common/BottomContent.kt
@@ -1,0 +1,69 @@
+/*
+ * Designed and developed by Duckie Team, 2022
+ *
+ * Licensed under the MIT.
+ * Please see full license: https://github.com/duckie-team/duckie-android/blob/develop/LICENSE
+ */
+
+package team.duckie.app.android.feature.ui.detail.common
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.DpSize
+import androidx.compose.ui.unit.dp
+import team.duckie.app.android.feature.ui.detail.viewmodel.state.DetailState
+import team.duckie.app.android.shared.ui.compose.Spacer
+import team.duckie.quackquack.ui.R
+import team.duckie.quackquack.ui.component.QuackDivider
+import team.duckie.quackquack.ui.component.QuackImage
+import team.duckie.quackquack.ui.component.QuackSmallButton
+import team.duckie.quackquack.ui.component.QuackSmallButtonType
+
+/** 상세 화면 최하단 Layout */
+@Composable
+internal fun DetailBottomLayout(
+    modifier: Modifier,
+    state: DetailState.Success,
+    onHeartClick: () -> Unit,
+    onChallengeClick: () -> Unit,
+) {
+    Column(modifier = modifier) {
+        Spacer(space = 20.dp)
+        // 구분선
+        QuackDivider()
+        // 버튼 모음 Layout
+        // TODO(riflockle7): 추후 Layout 을 활용해 처리하기
+        Row(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(horizontal = 16.dp, vertical = 9.dp),
+            verticalAlignment = Alignment.CenterVertically,
+            horizontalArrangement = Arrangement.SpaceBetween,
+        ) {
+            // 좋아요 버튼
+            QuackImage(
+                src = if (state.isHeart) {
+                    R.drawable.quack_ic_heart_filled_24
+                } else {
+                    R.drawable.quack_ic_heart_24
+                },
+                size = DpSize(24.dp, 24.dp),
+                onClick = onHeartClick,
+            )
+
+            // 버튼
+            QuackSmallButton(
+                text = state.buttonTitle,
+                type = QuackSmallButtonType.Fill,
+                enabled = true,
+                onClick = onChallengeClick,
+            )
+        }
+    }
+}
